Add files via upload
A detailed example where allowing multiple region states per location q in a tuple. The example tries to check if this solution is complete and indeed so. THe idea is whenever a new region state , say rq' , on q is encountered in a tuple that already contains a region state on q,  say rq : we check if rq' enables a new transition that rq doesnt. If so we add rq' , else we dont add rq' to tuple. Althoug this check may seem to be local (checking for transiton in q), as this exmple demonstrates it also takes care of transiton further down the line(\tau_7 in the isexample is only enable if rq' obtained via cycle on q2 and q is added to tuple. Although its not added in one path ther eis another path where ti is added and therfore enableing \tau_7)